Possible Dupe Spawn Fix: CreatureSpawns-Subnautica.cs
I looked in the code and all the reports. I noticed all the main creature dupe reports were of the ones with count before the probability. I will admit I am no programmer however logically I believe that the probability of the spawns checks for each count. while having count above probability it checks count first so an an example. count is 12 and probability is .01f idk the f I assume its % so a 1% chance of spawning 12 of the creature. while having probability first would seem to be like a 1% chance of spawning 1 of them and that triggers 12 times. idk makes sense to me could be complete nonsense. I will use my updated file and see if I get the dupe bug. I haven't played with it before. it will be a clean save and Ill give a update if it happens, and if u dont hear from me chances it worked.